requirements.txt
flask
motor
httpx
python-dotenv